+++ CMakeLists.txt
@@ -1,20 +1,21 @@
 cmake_minimum_required(VERSION 3.4.3)
+project(SVF-TOOL) # 设置项目名称
 if (DEFINED LLVM_DIR)
-        set(ENV{LLVM_DIR} "${LLVM_DIR}")
+    set(ENV{LLVM_DIR} "${LLVM_DIR}")
+endif()
+if (DEFINED ENV{LLVM_DIR})
+    # We need to match the build environment for LLVM:
+    # In particular, we need C++11 and the -fno-rtti flag
+    set(CMAKE_CXX_STANDARD 17)
+    if(CMAKE_BUILD_TYPE MATCHES "Debug")
+        set(CMAKE_CXX_FLAGS "-fPIC -std=gnu++14 -O0 -fno-rtti -Wno-deprecated")
+    else()
+        set(CMAKE_CXX_FLAGS "-fPIC -std=gnu++14 -O3 -fno-rtti -Wno-deprecated")
     endif()
-    if (DEFINED ENV{LLVM_DIR})
-        # We need to match the build environment for LLVM:
-        # In particular, we need C++11 and the -fno-rtti flag
-        set(CMAKE_CXX_STANDARD 14)
-	if(CMAKE_BUILD_TYPE MATCHES "Debug")
-		set(CMAKE_CXX_FLAGS "-fPIC -std=gnu++14 -O0 -fno-rtti -Wno-deprecated")
-	else()
-		set(CMAKE_CXX_FLAGS "-fPIC -std=gnu++14 -O3 -fno-rtti -Wno-deprecated")
-	endif()
-	set(CMAKE_C_FLAGS "-fPIC")
+    set(CMAKE_C_FLAGS "-fPIC")
 
-find_package(LLVM REQUIRED CONFIG)
-    
+    find_package(LLVM REQUIRED CONFIG)
+
     list(APPEND CMAKE_MODULE_PATH "${LLVM_CMAKE_DIR}")
     include(AddLLVM)
 
@@ -27,33 +28,37 @@ WARNING: The LLVM_DIR var was not set (required for an out-of-source build)!\n\
 Please set this to environment variable to point to the LLVM build directory\
 (e.g. on linux: export LLVM_DIR=/path/to/llvm/build/dir)")
 endif()
+set(SVF_DIR $ENV{SVF_DIR})
+message(${SVF_DIR})
 
-if (EXISTS "${SVF_DIR}")
+if(EXISTS "${SVF_DIR}")
 else()
-    set(SVF_DIR $ENV{SVF_DIR})
-    if(EXISTS "${SVF_DIR}")
-    else()
     message(FATAL_ERROR "\
 WARNING: The SVF_DIR var was not set (required for an out-of-source build)!\n\
 Please set this to environment variable to point to the SVF_DIR directory or set this variable to cmake configuration\n
 (e.g. on linux: export SVF_DIR=/path/to/SVF/dir) \n or \n \n(make the project via: cmake -DSVF_DIR=your_path_to_SVF) ")
-    endif()
 endif()
 
 if(CMAKE_BUILD_TYPE MATCHES "Debug")
     MESSAGE (STATUS "building SVF in debug mode")
     if (EXISTS "${SVF_DIR}/Debug-build")
-        set(SVF_LIB "${SVF_DIR}/Debug-build/lib/libSvf.a")
+        set(SVF_BIN "${SVF_DIR}/Debug-build")
     else()
-        set(SVF_LIB "${SVF_DIR}/Release-build/lib/libSvf.a")
+        set(SVF_BIN "${SVF_DIR}/Release-build")
     endif()
 else()
     MESSAGE (STATUS "building SVF in release mode")
-    set(SVF_LIB "${SVF_DIR}/Release-build/lib/libSvf.a")
-
+    set(SVF_BIN "${SVF_DIR}/Release-build")
 endif()
-set(SVF_HEADER "${SVF_DIR}/include")
-include_directories(${SVF_HEADER})
+set(SVF_HEADER "${SVF_DIR}/svf/include")
+message(${SVF_HEADER})
+set(S)
+set(SVF_LLVM_HEADER "${SVF_DIR}/svf-llvm/include")
+set(SVF_LIB "${SVF_BIN}/svf-llvm/libSvfLLVM.a" "${SVF_BIN}/svf/libSvfCore.a")
+set(SVF_BIN_HEADER "${SVF_BIN}/include")
+include_directories(${SVF_HEADER}
+        ${SVF_LLVM_HEADER}
+        ${SVF_BIN_HEADER})
 
 #set z3 env
 if (DEFINED Z3_DIR)
@@ -91,4 +96,3 @@ else()
 endif()
 
 add_subdirectory(src/svfModule)
-add_subdirectory(src/svfLLVMUtil)
