+++ env.sh
@@ -2,7 +2,7 @@
 
 PROJECTHOME=$(pwd)
 sysOS=`uname -s`
-LLVMHome="llvm-13.0.0.obj"
+LLVMHome="llvm-14.0.0.obj"
 Z3Home="z3.obj"
 install_path=`npm root`
 export LLVM_DIR=$install_path/$LLVMHome
@@ -14,10 +14,10 @@ echo "export Z3_DIR=$install_path/$Z3Home" >> ~/.bashrc
 echo "export PATH=$LLVM_DIR/bin:$PROJECTHOME/bin:$PATH" >> ~/.bashrc
 if [[ $sysOS == "Darwin" ]]
 then 
-    export SVF_DIR=$install_path/SVF/
+    export SVF_DIR=$install_path/SVF
 elif [[ $sysOS == "Linux" ]]
 then 
-    export SVF_DIR=$install_path/SVF/
+    export SVF_DIR=$install_path/SVF
 fi 
 
 echo "LLVM_DIR="$LLVM_DIR
