+++ src/svfpy/Module/CL.py
@@ -1,0 +1,14 @@
+from .package import libSvfModule
+
+libSvfModule = libSvfModule
+
+
+class CL(object):
+    global libSvfModule
+    _cl = libSvfModule
+
+    def ParseCommandLineOptions(self):
+        self._cl.ParseCommandLineOptions()
+
+
+CL = CL()
