+++ src/svfpy/Module/SVFG.py
@@ -1,0 +1,13 @@
+from .package import libSvfModule
+libSvfModule = libSvfModule
+
+
+class SVFG(object):
+    global libSvfModule
+    svfg = libSvfModule
+
+    def delete(self):
+        self.svfg.deleteSvfg()
+
+
+SVFG = SVFG()
