+++ src/svfpy/Module/SVFIRBuilder.py
@@ -1,0 +1,14 @@
+from .package import libSvfModule
+from .SVFIR import SVFIR
+libSvfModule = libSvfModule
+
+
+class SVFIRBuilder(object):
+    global libSvfModule
+    svfIRBuilder = libSvfModule
+
+    def build(self):
+        self.svfIRBuilder.build()
+        return SVFIR
+
+SVFIRBuilder = SVFIRBuilder()
