+++ src/svfpy/Module/SVFModule.py
@@ -1,0 +1,14 @@
+from .package import libSvfModule
+
+libSvfModule = libSvfModule
+
+
+class SVFModule(object):
+    global libSvfModule
+    svfModule = libSvfModule
+
+    def buildSymbolTableInfo(self):
+        self.svfModule.buildSymbolTableInfo()
+
+
+SVFModule = SVFModule()
